Updated listShift.py
Right now, if you enter a shift value that is not an integer, it allows it but the function will say that you have entered something incorrect. By uncommenting (is that a word?) lines 25, 27, and 28, and indenting line 26, it will loop your input till you enter something that is allowed.